removed "+"/added tag
The "+" tag was handy for indicating recently added views/projects,
but it sets up a weird cycle when closing views. E.g., open a view,
close the view, then open it again. The tag cycles from "+" to "-"
to blank. I'd rather surprise the user with views appearing
outright than confusing them with the tag.